nextTurn refactoring

+++ module/condition/cautious.js
@@ -18,13 +18,14 @@ export async function createCondition(token, options = {}) {
 
     const ON_CREATE_MACRO = `
 const token = canvas.tokens.get('${token.id}');
+if (!token) return;
 await token.deleteAllMoraleConditions(game.hm3.Condition.CAUTIOUS);
 const unconscious = token.hasCondition(game.hm3.Condition.UNCONSCIOUS);
 if (!unconscious) {
     await game.hm3.Gm2GmSays("<b>" + token.name + "</b> is now <b>Cautious</b>, and will not Engage, must choose <b>Pass</b> if engaged, and cannot select the Counterstrike defense.", "Combat 16");
     if (token.isEngaged()) {
-        await game.combats.active.nextTurn(500); // delay so that other hooks are executed first
         await game.hm3.GmSays("<b>" + token.name + "</b> is <b>Cautious</b>, and must choose <b>Pass</b> if engaged. <b>Turn ends.</b>", "Combat 16");
+        token.turnEnds();
     }
 }`;
 
@@ -32,11 +33,12 @@ if (!unconscious) {
         ? ''
         : `
 const token = canvas.tokens.get('${token.id}');
+if (!token) return;
 const unconscious = token.hasCondition(game.hm3.Condition.UNCONSCIOUS);
 if (!unconscious) {
     await game.hm3.Gm2GmSays("<b>" + token.name + "</b> is still <b>Cautious</b>, and will not Engage, must choose <b>Pass</b> if engaged, and cannot select the Counterstrike defense.", "Combat 16");
     if (token.isEngaged()) {
-        await game.combats.active.nextTurn(500); // delay so that other hooks are executed first
+        token.turnEnds();
         await game.hm3.GmSays("<b>" + token.name + "</b> is <b>Cautious</b>, and must choose <b>Pass</b> if engaged. <b>Turn ends.</b>", "Combat 16");
     }
 }`;
